Split out all getstarted/prep stuff out of Home
