+++ app/components/Home.js
@@ -30,21 +30,11 @@ class Home extends Component{
   }
 
   static propTypes = {
-    address: PropTypes.string.isRequired,
-    port: PropTypes.string.isRequired,
-    passphrase: PropTypes.string.isRequired,
     isLoggedIn: PropTypes.bool.isRequired,
-    isLoggingIn: PropTypes.bool.isRequired,
     client: PropTypes.object,
-    error: PropTypes.string,
 
     getBalanceRequestAttempt: PropTypes.bool.isRequired,
     getStakeInfoRequestAttempt: PropTypes.bool.isRequired,
-
-    getLoaderRequestAttempt: PropTypes.bool.isRequired,
-    walletCreateRequestAttempt: PropTypes.bool.isRequired,
-    walletExistRequestAttempt: PropTypes.bool.isRequired,
-    walletOpenRequestAttempt: PropTypes.bool.isRequired,
   }
 
   handleBalanceClick = () => {
@@ -52,42 +42,11 @@ class Home extends Component{
   }
 
   render() {
-    const { address, port } = this.props;
-    const { isLoggedIn, isLoggingIn, client, error} = this.props;
+    const { isLoggedIn, client} = this.props;
 
     const { getBalanceRequestAttempt, getBalanceResponse } = this.props;
     const { getStakeInfoRequestAttempt, getStakeInfoResponse } = this.props;
-
-    const { loader, getLoaderRequestAttempt, getLoaderError, loaderRequest } = this.props;
-    const { walletCreateResponse, walletCreateRequestAttempt, walletCreateError } = this.props;
-    const { walletOpenResponse, walletOpenRequestAttempt, walletOpenError } = this.props;
-    const { walletExistResponse, walletExistRequestAttempt, walletExistError } = this.props;
-    const { walletCloseResponse, walletCloseRequestAttempt, walletCloseError } = this.props;
-    const { startRpcResponse, startRpcRequestAttempt, startRpcError } = this.props;
-    const { loadActiveDataFiltersAttempt } = this.props;
-    
-    /*  View that will be seen on fresh starts */
-    const getStarted = (
-      <div>
-        <p>{error}</p>
-        <h3>Welcome to Decrediton</h3>
-        <h5>Please enter the information below to connect to you dcrwallet</h5>
-        <LoginForm />
-      </div>);
-
-    /*  View that will be when logging in is occuring */
-    const getStartedLoggingIn = (
-      <div >
-        Logging in
-        <LinearProgress mode="indeterminate" />
-      </div>);
-
-    /* View that will be shown when an error on logging in occured */
-    const getStartedError = (
-      <div >
-        <p> {error} </p>
-      </div>);
-
+  
     /* View that will be seen when user has a set Client */
     const homeView = (
       <div >
@@ -109,109 +68,20 @@ class Home extends Component{
           onClick={() => this.props.loadActiveDataFiltersAttempt()}
           label='Load Active Data Filters'/>
       </div>);
-
-    const getStartedCreateWallet = (
+    const errorView = (
       <div>
-        <CreateWalletForm />
+        <p>Something went wrong</p>
       </div>);
 
-    const getStartedWalletCreating = (
-      <div >
-        <h3> Creating wallet </h3>
-        <LinearProgress mode="indeterminate" />
-      </div>);
-
-    const getStartedWalletLoader = (
-      <div >
-        <p>{error}</p>
-        <RaisedButton type="submit"
-          style={styles.buttons}
-          primary={true}
-          onClick={() => {loaderRequest(address, port);}}
-          label='Get Started'/>
-      </div>);
-
-    const getStartedGettingLoader = (
-      <div >
-        <h3>Getting wallet loader service</h3>
-        <LinearProgress mode="indeterminate" />
-      </div>);
-
-    const getStartedWalletExistRequest = (
-      <div >
-        <h3>Checking if wallet exists</h3>
-        <LinearProgress mode="indeterminate" />
-      </div>);
-
-    const getStartedWalletExist = (
-      <div>
-        <p>{error}</p>
-        <h3>Check if wallet exists</h3>
-      </div>);
-
-    const getStartedWalletOpen = (
-      <div>
-        <p >{error}</p>
-        <h3>Opening wallet</h3>
-        <h5>Please enter the information below to connect to you dcrwallet</h5>
-        <WalletOpenForm />
-      </div>);
-
-    const getStartedOpeningWallet = (
-      <div>
-        <h3>Opening wallet</h3>
-        <LinearProgress mode="indeterminate" />
-      </div>);
-
-    // Step 4 complete
     if (isLoggedIn) {
       if (client === undefined) {
-        return(getStarted);
+        return(errorView);
       } else {
         return(homeView);
       }
+    } else {
+      return(errorView);
     }
-    // Step 4 action
-    if (isLoggingIn) {
-      return (getStartedLoggingIn);
-    }
-    // Step 3 complete/ Step 4 start
-    if (walletOpenResponse !== null) {
-      return(getStarted);
-    }
-    // Step 3 action
-    if (walletOpenRequestAttempt) {
-      return (getStartedOpeningWallet);
-    }
-    // Step 2 complete/ Step 3 start
-    if (walletExistResponse !== null && walletExistResponse.exists) {
-      return(getStartedWalletOpen);
-    }
-    // Step 2b creating wallet
-    if (walletCreateRequestAttempt) {
-      return(getStartedWalletCreating);
-    }
-    // Step 2 wallet exist action complete, though
-    // wallet does not exist
-
-    if (walletExistResponse !== null && !walletExistResponse.exists) {
-      return(getStartedCreateWallet);
-    }
-    // Step 2 action
-    if (walletExistRequestAttempt) {
-      return(getStartedWalletExistRequest);
-    }
-    // Step 1 complete/ Step 2 start
-    if (loader !== null) {
-      return(getStartedWalletExist);
-    }
-
-    // Step 1 action
-    if (getLoaderRequestAttempt) {
-      return (getStartedGettingLoader);
-    }
-    // Step 1 start
-    return (getStartedWalletLoader);
   }
 }
