+++ app/containers/HomePage.js
@@ -9,33 +9,13 @@ import * as ControlActions from '../actions/ControlActions';
 
 function mapStateToProps(state) {
   return {
-    address: state.login.address,
-    port: state.login.port,
-    passphrase: state.login.passphrase,
     isLoggedIn: state.login.isLoggedIn,
-    isLoggingIn: state.login.isLoggingIn,
     client: state.login.client,
-    error: state.login.error,
-
-    // WalletServices
+    
     getBalanceRequestAttempt: state.grpc.getBalanceRequestAttempt,
     getBalanceResponse: state.grpc.getBalanceResponse,
     getStakeInfoRequestAttempt: state.grpc.getStakeInfoRequestAttempt,
     getStakeInfoResponse: state.grpc.getStakeInfoResponse,
-
-    // WalletLoaderServices
-    walletCreateRequestAttempt: state.walletLoader.walletCreateRequestAttempt,
-    walletCreateResponse: state.walletLoader.walletCreateResponse,
-    walletExistRequestAttempt: state.walletLoader.walletExistRequestAttempt,
-    walletExistResponse: state.walletLoader.walletExistResponse,
-
-    getLoaderRequestAttempt: state.walletLoader.getLoaderRequestAttempt,
-    loader: state.walletLoader.loader,
-
-    walletOpenRequestAttempt: state.walletLoader.walletOpenRequestAttempt,
-    walletOpenResponse: state.walletLoader.walletOpenResponse,
-
-
   };
 }
