+++ app/routes.js
@@ -6,6 +6,7 @@ import HomePage from './containers/HomePage';
 import HistoryPage from './containers/HistoryPage';
 import SendPage from './containers/SendPage';
 import ReceivePage from './containers/ReceivePage';
+import GetStartedPage from './containers/GetStartedPage';
 
 export default (
   <Route path="/" component={App}>
@@ -13,5 +14,7 @@ export default (
     <Route path="/history" component={HistoryPage} />
     <Route path="/send" component={SendPage} />
     <Route path="/receive" component={ReceivePage} />
+    <Route path="/getstarted" component={GetStartedPage} />
+    
   </Route>
 );
